+++ src/invoices/InvoiceForm/FieldBatchGroup/FieldBatchGroup.js
@@ -1,0 +1,41 @@
+import React, { useMemo } from 'react';
+import PropTypes from 'prop-types';
+
+import {
+  FieldSelection,
+  validateRequired,
+} from '@folio/stripes-acq-components';
+
+import { getBatchGroupsOptions } from '../../../common/utils';
+
+const FieldBatchGroup = ({ batchGroups, isEditPostApproval }) => {
+  const batchGroupsOptions = useMemo(() => {
+    return getBatchGroupsOptions(batchGroups);
+  }, [batchGroups]);
+
+  const labelId = 'ui-invoice.invoice.details.information.batchGroup';
+  const isDisabled = isEditPostApproval || batchGroups.length === 1;
+
+  return (
+    <FieldSelection
+      dataOptions={batchGroupsOptions}
+      disabled={isDisabled}
+      id="invoice-batch-groups"
+      labelId={labelId}
+      name="batchGroupId"
+      required
+      validate={validateRequired}
+    />
+  );
+};
+
+FieldBatchGroup.propTypes = {
+  batchGroups: PropTypes.arrayOf(PropTypes.object),
+  isEditPostApproval: PropTypes.bool,
+};
+
+FieldBatchGroup.defaultProps = {
+  isEditPostApproval: false,
+};
+
+export default FieldBatchGroup;