+++ test/bigtest/tests/InvoiceCreate/InvoiceCreate-test.js
@@ -14,6 +14,7 @@ describe('Invoice create', () => {
 
   beforeEach(async function () {
     this.server.createList('vendor', 2);
+    this.server.create('batchgroup');
     this.visit('/invoice/create');
     await invoiceForm.whenLoaded();
   });